Feature multiple employee found and naming function (#1)
* fix: alt shift f

* feat: added naming function

* Feat: Added format date function

Function for HelloID DateTime to AEOS DateTime

* Fix: Update leaveDate on create account

This is needed for reboarding of old employees. AEOS does not accept a date as ArivalDate if the LeaveDate is older.

* Feat: Moved update on correlate to config

* Feat: Multiple employees found same externalID

Script now only gives error with no information. This throw helps the HelloID administrator to know where to solve the problem

* Feat: ExportData added

Display the AEOSid (unique AEOS number in database): aref

* Fix: updated readme for UpdateOnCorrelate toggle

* Fix: finetune code to make it better readable

* Fix: updated forum link to connector post